requirements.txt
openai
streamlit
faiss-cpu
sentence-transformers
PyPDF2
Groq